requirements.txt
python-dotenv==1.0.1
pycryptodome==3.20.0
python-box==7.1.1
gql==3.5.0
requests_toolbelt==1.0.0
eth-account==0.12.2
pandas==2.2.2
diskcache==5.6.3
pre-commit==3.7.1
black
ruff
codespell
mypy
nbqa
types-requests
pytest
